move cli classes into cli package
